added made pag_id_width a 32 bit integer

+++ inc/tupleindexer/common/page_access_specification.h
@@ -8,7 +8,7 @@ typedef struct page_access_specs page_access_specs;
 struct page_access_specs
 {
 	// page_id_width is bytes required for storing page_id, it can be in range [1, 8] both inclusive
-	uint8_t page_id_width;
+	uint32_t page_id_width;
 
 	// size of each page that can be accessed
 	uint32_t page_size;
@@ -47,7 +47,7 @@ struct page_access_specs
 };
 
 // initialize all attributed of page_access_specs
-int initialize_page_access_specs(page_access_specs* pas_p, uint8_t page_id_width, uint32_t page_size, uint64_t NULL_PAGE_ID);
+int initialize_page_access_specs(page_access_specs* pas_p, uint32_t page_id_width, uint32_t page_size, uint64_t NULL_PAGE_ID);
 
 // if this function returns 0, that implies that the pas's passed params are invalid
 int is_valid_page_access_specs(const page_access_specs* pas_p);
